Add display name to runner model
This commit adds a method to the runner model that will return a display
name for the runner. It will return the runner's token if a runner
doesn't have a description or if it's empty. It will return the
description otherwise. This is in order to give each particular runner a
more friendly name when viewing them on a particular project. This
commit also updates the runner_projects index to use this display name
instead of using the token. And finally it adds spec tests to cover the
new method added.