Change gravatar_icon helper to handle ssl
This commit refactors the gravatar_icon helper method to handle ssl.
This is because when running gitlab-ci with ssl enabled, it gives a
warning because portions of the page are not secure. In addition to
refactoring this method, it also adds some configuration values to the
application.yml.example file that are needed for the gravatar_icon
helper method. It also adds default values for settings relating to
gitlab_ci and gravatar. This is because the values will need defaults
if they aren't specified in the application.yml. It also adds the
no_avatar.png since it is being used in the gravatar_icon helper
method. Finally, this commit adds spec tests to cover the
refactored method.