Do not use image's file name as alternative text
`image_tag` by default uses the image's file name for the img tag's alt
attribute. This is not preferable in many cases. E.g. it clutters the UI
with a long hash string if a gravatar could not be loaded.

To fix this, the `alt` option is set for `image_tag` calls.